+++ docs/api/apiv3/components/examples/program_body.yml
@@ -1,7 +1,7 @@
-# Example: Project request payload
+# Example: Program request payload
 ---
 value:
-  name: 'Project example'
+  name: 'Program example'
   statusExplanation:
     raw: Everything **fine**
   description:
@@ -9,6 +9,6 @@ value:
   customField123: 123
   _links:
     parent:
-      href: "/api/v3/projects/123"
+      href: "/api/v3/programs/123"
     customField456:
-      href: "/api/v3/users/315"
+      href: "/api/v3/programs/315"
