+++ docs/api/apiv3/paths/portfolio_form.yml
@@ -1,0 +1,53 @@
+# /api/v3/portfolios/{id}/form
+---
+post:
+  parameters:
+    - description: Portfolio id
+      example: '1'
+      in: path
+      name: id
+      required: true
+      schema:
+        type: integer
+  requestBody:
+    content:
+      application/json:
+        schema:
+          type: object
+        examples:
+          "with custom fields":
+            "$ref": "../components/examples/portfolio_body.yml"
+          "empty":
+            description: Empty request to get the form initially in order to start the guided update of a portfolio
+            value: {}
+  responses:
+    '200':
+      description: OK
+      headers: {}
+    '400':
+      $ref: "../components/responses/invalid_request_body.yml"
+    '403':
+      content:
+        application/hal+json:
+          schema:
+            $ref: "../components/schemas/error_response.yml"
+          examples:
+            response:
+              value:
+                _type: Error
+                errorIdentifier: urn:openproject-org:api:v3:errors:MissingPermission
+                message: You are not authorized to access this resource.
+      description: |-
+        Returned if the client does not have sufficient permissions.
+
+        **Required permission:** edit workspace in the portfolio
+      headers: {}
+    '406':
+      $ref: "../components/responses/missing_content_type.yml"
+    '415':
+      $ref: "../components/responses/unsupported_media_type.yml"
+  tags:
+    - Portfolios
+  description: ''
+  operationId: Portfolio_update_form
+  summary: Portfolio update form
